Django Checkbook Challenge Submission Assignment
A Django web site that displays a unified web site in look and style.  From the index page you can choose to create an account or to view a current account.  Once an account is created the data is saved in an SQL database which has the initial deposit and user information.  From that transaction html file you can make deposit and and withdrawals adding and subtracting from a saved balance by the SQL database.  Can also go strait to a transaction from the index page via a drop down option menu from the navbar.